requirements.txt
honeybee-core==1.42.3
dragonfly-schema==1.6.32;python_version>='3.7'
